+++ scripts/effect-engine/effect-engine.js
@@ -3,6 +3,7 @@ import { conditionComponent } from "./components/condition-component.js";
 import { modifierComponent } from "./components/modifier-component.js";
 import { persistentDamageComponent } from "./components/persistent-damage-component.js";
 import { resistanceComponent } from "./components/resistance-component.js";
+import { weaknessComponent } from "./components/weakness-component.js";
 
 let initialized = false;
 
@@ -13,6 +14,7 @@ export function initializeEffectEngine() {
   componentRegistry.register(modifierComponent);
   componentRegistry.register(persistentDamageComponent);
   componentRegistry.register(resistanceComponent);
+  componentRegistry.register(weaknessComponent);
 
   initialized = true;
 }